# docs/swagger.yaml
basePath: /api
...
host: localhost:8080
info:
  contact: {}
  description: REST API для книжного издательства (лаб. 4). Swagger UI на том же host:port,
    что и API, отправляет обычные браузерные куки — после входа остаётся auth_token,
    и запросы идут как у авторизованного пользователя. Чтобы проверить сценарий гостя,
    удалите куки для этого origin или откройте приватное окно.
  title: Publishing Backend API
  version: "2.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Проверяет логин/пароль, создаёт сессию в Redis, устанавливает куку
        session_id
      parameters:
      - description: login, password
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Аутентификация
      tags:
      - auth
...
  /auth/register:
    post:
      consumes:
      - application/json
      description: Создаёт нового пользователя с хэшированным паролем
      parameters:
      - description: login, password, name, role
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Регистрация
      tags:
      - auth
  /auth/profile:
    put:
      consumes:
      - application/json
      description: Изменяет имя и/или пароль текущего пользователя
      parameters:
      - description: name, password
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - CookieAuth: []
      summary: Обновить профиль
      tags:
      - auth
...
securityDefinitions:
  CookieAuth:
    in: cookie
    name: auth_token
    type: apiKey
swagger: "2.0"